pyproject.toml
[project]
name = "jirani-offline-library"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "dotenv>=0.9.9",
    "fastapi>=0.118.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.11.0",
    "pymupdf>=1.26.7",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
